move share project to shared folder
